#4 型定義の整理
