+++ src/hooks/useRecipes.ts
@@ -1,4 +1,4 @@
-import { useState, useEffect, useCallback } from 'react';
+import { useState, useEffect } from 'react';
 import { Recipe } from '@/types';
 import { useGemini } from './useGemini';
